#28 #87: update pom.xml and refactor project structure
